# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Minidrama API
  version: 0.0.0
  description: |
    Wave 1 skeleton contract. This document is the single source of truth for the HTTP surface:
    the server validates against it and the client types are generated from it, so the two cannot
    drift (`docs/architecture/tech-stack.md` T13).

    Only the endpoints that exist in the skeleton are described here. The full surface is designed
    in `docs/12-api-contracts.md` and `docs/design/api-contracts.md`; endpoints move into this file
    as they are implemented, so a path in this document always has a running handler.

    **Cross-origin access applies to every path below.** The client bundle is served by TikTok, so
    the API is never the page's own origin and every browser call is a cross-origin call. A request
    carrying an `Origin` header that is not on the server's configured allowlist is refused with
    `403 COMMON_ORIGIN_NOT_ALLOWED` before routing — the response is the standard error envelope
    and carries no `Access-Control-*` header — and the browser's `OPTIONS` preflight is answered by
    the server rather than by any operation described here. There is no wildcard: an unconfigured
    deployment allows no browser origin at all, which is an outage for the client and never a
    quietly open API. Credentials are never allowed; the session is the `Authorization` header the
    client attaches itself. `Idempotency-Key`, `Authorization` and `Content-Type` are the
...
servers:
...
paths:
...
  /v1/unlock/ad-sessions:
    post:
      operationId: createAdUnlockSession
      summary: Issue a nonce for one rewarded-ad show
      description: |
        Issues a single-use nonce for a permitted placement (`AFTER_EPISODE` or `MANUAL_SKIP` —
        the two slots F-4 allows). It does not unlock anything.

        The `adUnitId` is the server's configured Portal placement id. When none is configured
        (GATE-4 unanswered) the response is `503 UNLOCK_AD_UNAVAILABLE`, not a placeholder id.
        The client must not bake an id into the bundle.

        A nonce is valid for one show. Reporting `isEnded: false` consumes it.
      tags: [unlock]
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdUnlockSessionRequest'
      responses:
        '201':
          description: A nonce and the unit id to hand to `createRewardedVideoAd`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdUnlockSession'
        '400':
          description: |
            `COMMON_VALIDATION_FAILED`. `episodeId` missing, or `placement` is not one of the
            two permitted slots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: '`AUTH_REQUIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: '`CONTENT_NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: '`CONTENT_OFFLINE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            `UNLOCK_POLICY_NOT_ALLOWED` or `UNLOCK_ALREADY_UNLOCKED`. Free, VIP-only, and
            already-owned episodes are not an ad grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: |
            `UNLOCK_AD_UNAVAILABLE` when no rewarded placement id is configured.
            `COMMON_SERVICE_UNAVAILABLE` when entitlement facts cannot be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/unlock/ad-grants:
    post:
      operationId: createAdUnlockGrant
      summary: Grant an episode after a completed rewarded ad
      description: |
        The only write that turns a completed rewarded ad into an `AD` unlock receipt.

        The client reports `isEnded` from the SDK close callback. That boolean is not a grant:
        the server verifies it through a replaceable completion check (U-18: there is no
        platform SSV callback), consumes the nonce, enforces the daily quota, and writes the
        receipt. A 200 with `isEnded: true` and no nonce is still a refuse.

        `Idempotency-Key` is required. A repeat of the same key by the same viewer for the same
        episode returns the first result.

        Interstitials do not unlock and have no grant endpoint. Wallet top-up is unrelated and
        stays disabled.
      tags: [unlock]
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdUnlockGrantRequest'
      responses:
        '200':
          description: The `AD` receipt and today's quota.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdUnlockGrant'
        '400':
          description: |
            `COMMON_VALIDATION_FAILED` or `COMMON_IDEMPOTENCY_KEY_REQUIRED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: '`AUTH_REQUIRED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: '`UNLOCK_ALREADY_UNLOCKED` or `COMMON_IDEMPOTENCY_CONFLICT`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            `UNLOCK_AD_NOT_COMPLETED` when the verifier did not accept the view;
            `UNLOCK_POLICY_NOT_ALLOWED` when the nonce is unusable or the episode cannot take
            an ad grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: '`UNLOCK_AD_QUOTA_EXCEEDED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: '`COMMON_SERVICE_UNAVAILABLE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateAdUnlockSessionRequest:
      type: object
      required: [episodeId, placement]
      properties:
        episodeId:
          type: string
          minLength: 1
        placement:
          type: string
          enum: [AFTER_EPISODE, MANUAL_SKIP]
          description: |
            The two F-4 slots. An unlock-panel placement is not expressible.
    AdUnlockSession:
      type: object
      required: [nonce, adUnitId, placement, episodeId, expiresAt]
      properties:
        nonce:
          type: string
          description: Single-use. Reporting a skip consumes it.
        adUnitId:
          type: string
          description: Server-configured Portal id. Absent configuration is a 503, not this field.
        placement:
          type: string
          enum: [AFTER_EPISODE, MANUAL_SKIP]
        episodeId:
          type: string
        expiresAt:
          type: string
          format: date-time
    CreateAdUnlockGrantRequest:
      type: object
      required: [episodeId, nonce, isEnded]
      properties:
        episodeId:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
        isEnded:
          type: boolean
          description: |
            The SDK close callback. `true` is still not a grant: the server verifier, the nonce,
            and the quota all sit between this field and the unlock row.
    AdUnlockGrant:
      type: object
      required: [unlock, quota]
      properties:
        unlock:
          type: object
          required: [id, episodeId, method, costCoins]
          properties:
            id:
              type: string
            episodeId:
              type: string
            method:
              type: string
              const: AD
            costCoins:
              type: integer
              const: 0
        quota:
          type: object
          required: [usedToday, dailyLimit]
          properties:
            usedToday:
              type: integer
              minimum: 0
            dailyLimit:
              type: integer
              minimum: 0
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code:
              type: string
            message:
              type: string
            traceId:
              type: string
              description: Echoed so a user report maps to a server trace.
            details:
              type: [object, 'null']
              additionalProperties: true
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      description: |
        The opaque session token from `POST /v1/auth/login`, as `Authorization: Bearer <token>`. It
        is not a JWT and carries no claims: the mapping from token to user is the server's, so a
        token cannot be constructed. Declared on the endpoints that are per-viewer; endpoints without
        it either serve everyone the same answer or enforce their own rule.
...